+++ src/mcmc/external_sampler.jl
@@ -7,6 +7,23 @@ The `Unconstrained` type-parameter is to indicate whether the sampler requires u
 
 # Fields
 $(TYPEDFIELDS)
+
+# Turing.jl's interface for external samplers
+
+When implementing a new `MySampler <: AbstractSampler`,
+`MySampler` must first and foremost conform to the `AbstractMCMC` interface to work with Turing.jl's `externalsampler` function.
+In particular, it must implement:
+
+- `AbstractMCMC.step` (the main function for taking a step in MCMC sampling; this is documented in AbstractMCMC.jl)
+- `AbstractMCMC.getparams(::DynamicPPL.Model, external_state)`: How to extract the parameters from the state returned by your sampler (i.e., the second return value of `step`).
+
+There are a few more optional functions which you can implement to improve the integration with Turing.jl:
+
+- `Turing.Inference.isgibbscomponent(::MySampler)`: If you want your sampler to function as a component in Turing's Gibbs sampler, you should make this evaluate to `true`.
+
+- `Turing.Inference.requires_unconstrained_space(::MySampler)`: If your sampler requires unconstrained space, you should return `true`. This tells Turing to perform linking on the VarInfo before evaluation, and ensures that the parameter values passed to your sampler will always be in unconstrained (Euclidean) space.
+
+- `Turing.Inference.getlogp_external(external_transition, external_state)`: Tell Turing how to extract the log probability density associated with this transition (and state). If you do not specify these, Turing will simply re-evaluate the model with the parameters obtained from `getparams`, which can be inefficient. It is therefore recommended to store the log probability density in either the transition or the state (or both) and override this method.
 """
 struct ExternalSampler{S<:AbstractSampler,AD<:ADTypes.AbstractADType,Unconstrained} <:
        InferenceAlgorithm
@@ -68,30 +85,28 @@ function externalsampler(
     return ExternalSampler(sampler, adtype, Val(unconstrained))
 end
 
-struct TuringState{S,M,V,C}
+"""
+    getlogp_external(external_transition, external_state)
+
+Get the log probability density associated with the external sampler's
+transition and state. Returns `missing` by default; in this case, an extra
+model evaluation will be needed to calculate the correct log density.
+"""
+getlogp_external(::Any, ::Any) = missing
+getlogp_external(mh::AdvancedMH.Transition, ::AdvancedMH.Transition) = mh.lp
+getlogp_external(hmc::AdvancedHMC.Transition, ::AdvancedHMC.HMCState) = hmc.stat.log_density
+
+struct TuringState{S,V1<:AbstractVarInfo,M,V,C}
     state::S
+    # Note that this varinfo has the correct parameters and logp obtained from
+    # the state, whereas `ldf.varinfo` will in general have junk inside it.
+    varinfo::V1
     ldf::DynamicPPL.LogDensityFunction{M,V,C}
 end
 
-state_to_turing(f::DynamicPPL.LogDensityFunction, state) = TuringState(state, f)
-function transition_to_turing(f::DynamicPPL.LogDensityFunction, transition)
-    # TODO: We should probably rename this `getparams` since it returns something
-    # very different from `Turing.Inference.getparams`.
-    θ = getparams(f.model, transition)
-    varinfo = DynamicPPL.unflatten(f.varinfo, θ)
-    return Transition(f.model, varinfo, transition)
-end
-
-function varinfo(state::TuringState)
-    θ = getparams(state.ldf.model, state.state)
-    # TODO: Do we need to link here first?
-    return DynamicPPL.unflatten(state.ldf.varinfo, θ)
-end
+varinfo(state::TuringState) = state.varinfo
 varinfo(state::AbstractVarInfo) = state
 
-# NOTE: Only thing that depends on the underlying sampler.
-# Something similar should be part of AbstractMCMC at some point:
-# https://github.com/TuringLang/AbstractMCMC.jl/pull/86
 getparams(::DynamicPPL.Model, transition::AdvancedHMC.Transition) = transition.z.θ
 function getparams(model::DynamicPPL.Model, state::AdvancedHMC.HMCState)
     return getparams(model, state.transition)
@@ -99,6 +114,21 @@ end
 getstats(transition::AdvancedHMC.Transition) = transition.stat
 
 getparams(::DynamicPPL.Model, transition::AdvancedMH.Transition) = transition.params
+
+function make_updated_varinfo(
+    f::DynamicPPL.LogDensityFunction, external_transition, external_state
+)
+    # Set the parameters.
+    new_parameters = getparams(f.model, external_state)
+    new_varinfo = DynamicPPL.unflatten(f.varinfo, new_parameters)
+    # Set (or recalculate, if needed) the log density.
+    new_logp = getlogp_external(external_transition, external_state)
+    return if ismissing(new_logp)
+        last(DynamicPPL.evaluate!!(f.model, new_varinfo, f.context))
+    else
+        DynamicPPL.setlogp!!(new_varinfo, new_logp)
+    end
+end
 
 # TODO: Do we also support `resume`, etc?
 function AbstractMCMC.step(
@@ -143,8 +173,15 @@ function AbstractMCMC.step(
             kwargs...,
         )
     end
+
+    # Get the parameters and log density, and set them in the varinfo.
+    new_varinfo = make_updated_varinfo(f, transition_inner, state_inner)
+
     # Update the `state`
-    return transition_to_turing(f, transition_inner), state_to_turing(f, state_inner)
+    return (
+        Transition(f.model, new_varinfo, transition_inner),
+        TuringState(state_inner, new_varinfo, f),
+    )
 end
 
 function AbstractMCMC.step(
@@ -157,11 +194,17 @@ function AbstractMCMC.step(
     sampler = sampler_wrapper.alg.sampler
     f = state.ldf
 
-    # Then just call `AdvancedHMC.step` with the right arguments.
+    # Then just call `AdvancedMCMC.step` with the right arguments.
     transition_inner, state_inner = AbstractMCMC.step(
         rng, AbstractMCMC.LogDensityModel(f), sampler, state.state; kwargs...
     )
 
+    # Get the parameters and log density, and set them in the varinfo.
+    new_varinfo = make_updated_varinfo(f, transition_inner, state_inner)
+
     # Update the `state`
-    return transition_to_turing(f, transition_inner), state_to_turing(f, state_inner)
+    return (
+        Transition(f.model, new_varinfo, transition_inner),
+        TuringState(state_inner, new_varinfo, f),
+    )
 end